requirements/prod.txt
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
babel==0.9.6
basket-client==0.3.7
billiard==2.7.3.34
bleach==1.1.5
boto==2.9.6
cef==0.5
celery==3.0.24
celery-tasktree==0.3.2
certifi==0.0.8
chardet==1.0.1
commonware==0.4.2
cssutils==0.9.7b3
curling==0.2.6
defusedxml==0.4.1
Django==1.4.8
dj-database-url==0.2.2
django-aesfield==0.1
django-browserid==0.8
django-cache-machine==0.8
django-celery==3.0.23
django-cronjobs==0.2.3
django_csp==1.0.2
django-extensions==1.2.1
django-filter==0.7
django-memcached-pool==0.4.1
django-multidb-router==0.5
django-mysql-pool==0.2
django-mysql-pymysql==0.1
django-nose==1.2
django-picklefield==0.2.1
django-quieter-formset==0.3
django-raven-heka==0.3
django-ratelimit==0.1
django-recaptcha-mozilla==0.0.1
djangorestframework==2.3.9
#django-session-csrf==0.6
django-statsd-mozilla==0.3.10
django-storages==1.1.8
django-waffle==0.9.1
easy-thumbnails==1.3
elasticutils==0.8.1
email-reply-parser==0.1.9
fastchardet==0.2.0
feedparser==5.1.3
fudge==1.0.3
GitPython==0.1.7
google-api-python-client==1.0
gunicorn==0.17.4
hera==0.1.2
html5lib==0.90
httplib2==0.8.0
importlib-no-failure==1.0.2
kombu==2.5.16
heka-py==0.30.3
heka-py-cef==0.3.1
heka-py-raven==0.6
python-mimeparse==0.1.4
mock==1.0.1
monolith.client==0.9
moz-addon-packager==1.0.1
mozpay==2.0.0
m2secret==0.1.1
ndg-httpsclient==0.3.2
nose==1.3.0
newrelic==2.4.0.4
oauth2==1.5.211
oauthlib==0.4.0
ordereddict==1.1
pyasn1==0.1.7
protobuf==2.5.0
PyBrowserID==0.6
pyelasticsearch==0.6
pyes==0.16.0
PyJWT-mozilla==0.1.4.2
PyMySQL==0.5
pyquery==0.4
python-dateutil==1.5
python-gflags==2.0
python-memcached==1.53
pytz==2013b
raven==3.1.16
rdflib==3.0.0
recaptcha-client==1.0.5
receipts==0.2.6
redis==2.7.6
requests==2.0.0
schematic==0.2
#signing_clients==0.1.7
six==1.3.0
slumber==0.5.3
SQLAlchemy==0.7.5
statsd==2.0.3
suds==0.3.9

## Not on pypi.
-e git+https://github.com/mozilla/amo-validator.git@40886c0705649d421e04eafd724877002e497007#egg=amo-validator
-e git+https://github.com/mozilla/app-validator.git@3725f5fe8e75e537c7f4b5392bb6eb8a7b3f2f15#egg=app-validator
-e git+https://github.com/jbalogh/django-queryset-transform@a1ba6ae41bd86f5bb9ff66fb56614e0fafe6e022#egg=django-queryset-transform
-e git+https://github.com/miracle2k/django-tables.git@546f339308103880c823b2056830fcdc9220edd0#egg=django-tables
-e git+https://github.com/mozilla/happyforms.git@729612c2a824a7e8283d416d2084bf506c671e24#egg=happyforms

# Temporary fork.
-e git+https://github.com/jsocol/jingo-minify.git@b7405d3f93628190bf83cc43b0ff44e2fbb8b3c0#egg=jingo_minify
-e git+https://github.com/mozilla/nuggets.git@96e80a64aa4bfcfef4f43fc3ab6966450ccd7325#egg=nuggets
-e git+https://github.com/jbalogh/test-utils.git@ce5136a257cd44a1c663319124a255c1d10a9834#egg=test-utils
-e git+https://github.com/fwenzel/django-mozilla-product-details.git@36ef06539d6b34c4f345fd0d3e16937d0db9a752#egg=django-mozilla-product-details
-e git+https://github.com/mozilla/signing-clients@a8bd730b202391c080113d224d223463e03088e9#egg=signing-clients
-e git+https://github.com/mozilla/django-session-csrf@f00ad913c62e139d36078e8a7e07dab65a021386#egg=django-session-csrf

## Forked.
-e git+https://github.com/andymckay/django-piston-oauth2.git@fa28ee#egg=django-piston-oauth2
-e git+https://github.com/kumar303/django-qunit.git@b0f468dcf33439488158c845df37ef3261852b55#egg=django-qunit
-e git+https://github.com/andymckay/django-uuidfield.git@029dd1263794ec36c327617cd6c2346da81c8c33#egg=django-uuidfield


## Out of date or not sure on pypi.
-e git+https://github.com/jbalogh/django-mobility.git@e2b60a1f96e4c4aed736395c01bf707e969d8e83#egg=django-mobility
-e git+https://github.com/jbalogh/jingo.git@b4ecca1066bcd1eac87358b4c82902442d4eedc4#egg=jingo
-e git+https://github.com/clouserw/tower.git@ce2eba049a5146918fa5239d827bdde9e36eadb7#egg=tower-dev
